requirements.txt
Django==3.2.16
pytest==6.2.4
pytest-pythonpath==0.7.3
pytest-django==4.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Pillow==9.3.0
PyJWT==2.1.0
requests==2.26.0
asgiref==3.7.2
attrs==23.1.0
black==22.1.0
blue==0.9.1
certifi==2023.11.17
charset-normalizer==2.0.12
click==8.1.7
Django==3.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-yasg==1.21.7
flake8==4.0.1
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
mccabe==0.6.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
Pillow==9.3.0
platformdirs==4.1.0
pluggy==0.13.1
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.26.0
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==1.26.18
asgiref==3.7.2
attrs==23.1.0
black==22.1.0
blue==0.9.1
certifi==2023.11.17
charset-normalizer==2.0.12
click==8.1.7
Django==3.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
flake8==4.0.1
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
mccabe==0.6.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
Pillow==9.3.0
platformdirs==4.1.0
pluggy==0.13.1
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.26.0
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==1.26.18
